Move distance functions to world plane class
