+++ src/utils/distance_utils.hpp
@@ -7,37 +7,15 @@ namespace rgbd_slam {
     namespace utils {
 
         /**
-         * \brief Compute a signed distance between two points
+         * \brief compute a distance bewteen two angles in radian
+         * \param[in] angleA
+         * \param[in] angleB
          */
-        vectorxd get_signed_distance(const vectorxd& pointA, const vectorxd& pointB);
-
-        /**
-         * \brief Compute a signed distance between a world point and a 3D point in screen space, by projecting the screen point to world space
-         * \param[in] worldPoint A 3D point in world space
-         * \param[in] screenPoint A 3D point in screen space
-         * \param[in] cameraToWorld A matrix to convert from camera to world space
-         * \return The 3D signed distance in world space
-         */
-        vector3 get_3D_to_3D_distance_3D(const WorldCoordinate& worldPoint, const ScreenCoordinate& screenPoint, const cameraToWorldMatrix& cameraToWorld);
-
-        /**
-         * \brief Compute a distance between a world point and a 3D point in screen space, by projecting the screen point to world space
-         * \param[in] worldPoint A 3D point in world space
-         * \param[in] screenPoint A 3D point in screen space
-         * \param[in] cameraToWorld A matrix to convert from camera to world space
-         * \return The unsigned distance in world space
-         */
-        double get_3D_to_3D_distance(const WorldCoordinate& worldPoint, const ScreenCoordinate& screenPoint, const cameraToWorldMatrix& cameraToWorld);
-
-        /**
-         * \brief Compute a distance between two planes, by retroprojecting a world plane to camera space
-         * \param[in] worldPlane A plane in world coordinates
-         * \param[in] cameraPlane A plane in camera coordinates
-         * \param[in] worldToCamera A transformation matrix to convert from world to camera space
-         *
-         * \return A 3D vector of the error between the two planes. The x and y are angle distances, the z is in millimeters
-         */
-        vector4 get_3D_to_2D_plane_distance(const PlaneWorldCoordinates& worldPlane, const PlaneCameraCoordinates& cameraPlane, const planeWorldToCameraMatrix& worldToCamera);
+        double angle_distance(const double angleA, const double angleB)
+        {
+            return atan2(sin(angleA - angleB), cos(angleA - angleB));
+        }
+        
 
     }   // utils
 }       // rgbd_slam
